Enable logging for cargo test
